backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
pgvector
python-jose[cryptography]
passlib[bcrypt]

bcrypt>=4.0.1
email-validator>=2.0.0
pytest-asyncio
google-generativeai
